+++ banyand/stream/block.go
@@ -439,21 +439,16 @@ func (bc *blockCursor) reset() {
 	bc.tagFamilies = tff[:0]
 }
 
-func initBlockCursor[T queryOptions | filterOptions](bc *blockCursor, p *part, bm *blockMetadata, opts T) {
+func (bc *blockCursor) init(p *part, bm *blockMetadata, opts queryOptions) {
 	bc.reset()
 	bc.p = p
 	bc.bm.copyFrom(bm)
-	switch opts := interface{}(opts).(type) {
-	case queryOptions:
-		bc.minTimestamp = opts.minTimestamp
-		bc.maxTimestamp = opts.maxTimestamp
-		bc.tagProjection = opts.TagProjection
-	case filterOptions:
-		bc.minTimestamp = opts.minTimestamp
-		bc.maxTimestamp = opts.maxTimestamp
+	bc.minTimestamp = opts.minTimestamp
+	bc.maxTimestamp = opts.maxTimestamp
+	bc.tagProjection = opts.TagProjection
+	if opts.elementRefMap != nil {
 		seriesID := bc.bm.seriesID
 		bc.expectedTimestamps = opts.elementRefMap[seriesID]
-		bc.tagProjection = opts.TagProjection
 	}
 }
 
@@ -536,56 +531,7 @@ func (bc *blockCursor) copyTo(r *pbv1.StreamResult) {
 	}
 }
 
-func (bc *blockCursor) loadData(tmpBlock *block) bool {
-	tmpBlock.reset()
-	bc.bm.tagProjection = bc.tagProjection
-	var tf map[string]*dataBlock
-	for i := range bc.tagProjection {
-		for tfName, block := range bc.bm.tagFamilies {
-			if bc.tagProjection[i].Family == tfName {
-				if tf == nil {
-					tf = make(map[string]*dataBlock, len(bc.tagProjection))
-				}
-				tf[tfName] = block
-			}
-		}
-	}
-	bc.bm.tagFamilies = tf
-	tmpBlock.mustReadFrom(&bc.tagValuesDecoder, bc.p, bc.bm)
-
-	start, end, ok := timestamp.FindRange(tmpBlock.timestamps, bc.minTimestamp, bc.maxTimestamp)
-	if !ok {
-		return false
-	}
-	bc.timestamps = append(bc.timestamps, tmpBlock.timestamps[start:end+1]...)
-	bc.elementIDs = append(bc.elementIDs, tmpBlock.elementIDs[start:end+1]...)
-
-	for i, projection := range bc.bm.tagProjection {
-		tf := tagFamily{
-			name: projection.Family,
-		}
-		blockIndex := 0
-		for _, name := range projection.Names {
-			t := tag{
-				name: name,
-			}
-			if len(tmpBlock.tagFamilies[i].tags) != 0 && tmpBlock.tagFamilies[i].tags[blockIndex].name == name {
-				t.valueType = tmpBlock.tagFamilies[i].tags[blockIndex].valueType
-				if len(tmpBlock.tagFamilies[i].tags[blockIndex].values) != len(tmpBlock.timestamps) {
-					logger.Panicf("unexpected number of values for tags %q: got %d; want %d",
-						tmpBlock.tagFamilies[i].tags[blockIndex].name, len(tmpBlock.tagFamilies[i].tags[blockIndex].values), len(tmpBlock.timestamps))
-				}
-				t.values = append(t.values, tmpBlock.tagFamilies[i].tags[blockIndex].values[start:end+1]...)
-			}
-			blockIndex++
-			tf.tags = append(tf.tags, t)
-		}
-		bc.tagFamilies = append(bc.tagFamilies, tf)
-	}
-	return true
-}
-
-func (bc *blockCursor) searchData(tmpBlock *block) bool {
+func (bc *blockCursor) loadData(tmpBlock *block, applyFilter bool) bool {
 	tmpBlock.reset()
 	bc.bm.tagProjection = bc.tagProjection
 	var tf map[string]*dataBlock
@@ -603,17 +549,28 @@ func (bc *blockCursor) searchData(tmpBlock *block) bool {
 	tmpBlock.mustReadFrom(&bc.tagValuesDecoder, bc.p, bc.bm)
 
 	idxList := make([]int, 0)
-	for _, ts := range bc.expectedTimestamps {
-		idx := timestamp.Find(tmpBlock.timestamps, ts)
-		if idx == -1 {
-			continue
+	var start, end int
+	if applyFilter {
+		for _, ts := range bc.expectedTimestamps {
+			idx := timestamp.Find(tmpBlock.timestamps, ts)
+			if idx == -1 {
+				continue
+			}
+			idxList = append(idxList, idx)
+			bc.timestamps = append(bc.timestamps, tmpBlock.timestamps[idx])
+			bc.elementIDs = append(bc.elementIDs, tmpBlock.elementIDs[idx])
 		}
-		idxList = append(idxList, idx)
-		bc.timestamps = append(bc.timestamps, tmpBlock.timestamps[idx])
-		bc.elementIDs = append(bc.elementIDs, tmpBlock.elementIDs[idx])
-	}
-	if len(bc.timestamps) == 0 {
-		return false
+		if len(bc.timestamps) == 0 {
+			return false
+		}
+	} else {
+		s, e, ok := timestamp.FindRange(tmpBlock.timestamps, bc.minTimestamp, bc.maxTimestamp)
+		start, end = s, e
+		if !ok {
+			return false
+		}
+		bc.timestamps = append(bc.timestamps, tmpBlock.timestamps[s:e+1]...)
+		bc.elementIDs = append(bc.elementIDs, tmpBlock.elementIDs[s:e+1]...)
 	}
 
 	for i, projection := range bc.bm.tagProjection {
@@ -631,8 +588,12 @@ func (bc *blockCursor) searchData(tmpBlock *block) bool {
 					logger.Panicf("unexpected number of values for tags %q: got %d; want %d",
 						tmpBlock.tagFamilies[i].tags[blockIndex].name, len(tmpBlock.tagFamilies[i].tags[blockIndex].values), len(tmpBlock.timestamps))
 				}
-				for _, idx := range idxList {
-					t.values = append(t.values, tmpBlock.tagFamilies[i].tags[blockIndex].values[idx])
+				if applyFilter {
+					for _, idx := range idxList {
+						t.values = append(t.values, tmpBlock.tagFamilies[i].tags[blockIndex].values[idx])
+					}
+				} else {
+					t.values = append(t.values, tmpBlock.tagFamilies[i].tags[blockIndex].values[start:end+1]...)
 				}
 			}
 			blockIndex++
