+++ banyand/stream/stream.go
@@ -60,9 +60,9 @@ type Stream interface {
 	io.Closer
 	GetSchema() *databasev1.Stream
 	GetIndexRules() []*databasev1.IndexRule
-	Query(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamResultPuller, error)
-	Sort(ctx context.Context, opts pbv1.StreamSortOptions) (pbv1.StreamSortResult, error)
-	Filter(ctx context.Context, opts pbv1.StreamFilterOptions) (pbv1.StreamResultPuller, error)
+	Query(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error)
+	Sort(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamSortResult, error)
+	Filter(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error)
 }
 
 var _ Stream = (*stream)(nil)
