--- banyand/stream/utils.go
@@ -1,169 +1,0 @@
-// Licensed to Apache Software Foundation (ASF) under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Apache Software Foundation (ASF) licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package stream
-
-import (
-	"github.com/apache/skywalking-banyandb/api/common"
-	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
-	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	"github.com/apache/skywalking-banyandb/pkg/convert"
-	"github.com/apache/skywalking-banyandb/pkg/logger"
-	"github.com/apache/skywalking-banyandb/pkg/partition"
-	pbv1 "github.com/apache/skywalking-banyandb/pkg/pb/v1"
-)
-
-func mustEncodeTagValue(name string, tagType databasev1.TagType, tagValue *modelv1.TagValue, num int) [][]byte {
-	values := make([][]byte, num)
-	nv := encodeTagValue(name, tagType, tagValue)
-	value := nv.marshal()
-	for i := 0; i < num; i++ {
-		values[i] = value
-	}
-	return values
-}
-
-func mustDecodeTagValue(valueType pbv1.ValueType, value []byte) *modelv1.TagValue {
-	if value == nil {
-		switch valueType {
-		case pbv1.ValueTypeInt64:
-			logger.Panicf("int64 can be nil")
-		case pbv1.ValueTypeStr:
-			return pbv1.EmptyStrTagValue
-		case pbv1.ValueTypeStrArr:
-			return pbv1.EmptyStrArrTagValue
-		case pbv1.ValueTypeInt64Arr:
-			return pbv1.EmptyIntArrTagValue
-		case pbv1.ValueTypeBinaryData:
-			return pbv1.EmptyBinaryTagValue
-		default:
-			return pbv1.NullTagValue
-		}
-	}
-	switch valueType {
-	case pbv1.ValueTypeInt64:
-		return int64TagValue(convert.BytesToInt64(value))
-	case pbv1.ValueTypeStr:
-		return strTagValue(string(value))
-	case pbv1.ValueTypeBinaryData:
-		return binaryDataTagValue(value)
-	case pbv1.ValueTypeInt64Arr:
-		var values []int64
-		for i := 0; i < len(value); i += 8 {
-			values = append(values, convert.BytesToInt64(value[i:i+8]))
-		}
-		return int64ArrTagValue(values)
-	case pbv1.ValueTypeStrArr:
-		var values []string
-		bb := bigValuePool.Generate()
-		var err error
-		for len(value) > 0 {
-			bb.Buf, value, err = unmarshalVarArray(bb.Buf[:0], value)
-			if err != nil {
-				logger.Panicf("unmarshalVarArray failed: %v", err)
-			}
-			values = append(values, string(bb.Buf))
-		}
-		return strArrTagValue(values)
-	default:
-		logger.Panicf("unsupported value type: %v", valueType)
-		return nil
-	}
-}
-
-func int64TagValue(value int64) *modelv1.TagValue {
-	return &modelv1.TagValue{
-		Value: &modelv1.TagValue_Int{
-			Int: &modelv1.Int{
-				Value: value,
-			},
-		},
-	}
-}
-
-func strTagValue(value string) *modelv1.TagValue {
-	return &modelv1.TagValue{
-		Value: &modelv1.TagValue_Str{
-			Str: &modelv1.Str{
-				Value: value,
-			},
-		},
-	}
-}
-
-func binaryDataTagValue(value []byte) *modelv1.TagValue {
-	data := make([]byte, len(value))
-	copy(data, value)
-	return &modelv1.TagValue{
-		Value: &modelv1.TagValue_BinaryData{
-			BinaryData: data,
-		},
-	}
-}
-
-func int64ArrTagValue(values []int64) *modelv1.TagValue {
-	return &modelv1.TagValue{
-		Value: &modelv1.TagValue_IntArray{
-			IntArray: &modelv1.IntArray{
-				Value: values,
-			},
-		},
-	}
-}
-
-func strArrTagValue(values []string) *modelv1.TagValue {
-	return &modelv1.TagValue{
-		Value: &modelv1.TagValue_StrArray{
-			StrArray: &modelv1.StrArray{
-				Value: values,
-			},
-		},
-	}
-}
-
-func (s *stream) genIndex(tagProj []pbv1.TagProjection, seriesList pbv1.SeriesList) (map[string]int, map[string]*databasev1.TagSpec,
-	map[string]partition.TagLocator, map[common.SeriesID]int,
-) {
-	entityMap := make(map[string]int)
-	for idx, entity := range s.schema.GetEntity().GetTagNames() {
-		entityMap[entity] = idx + 1
-	}
-	tagSpecIndex := make(map[string]*databasev1.TagSpec)
-	for _, tagFamilySpec := range s.schema.GetTagFamilies() {
-		for _, tagSpec := range tagFamilySpec.Tags {
-			tagSpecIndex[tagSpec.GetName()] = tagSpec
-		}
-	}
-	tagProjIndex := make(map[string]partition.TagLocator)
-	for i, tagFamilyProj := range tagProj {
-		for j, tagProj := range tagFamilyProj.Names {
-			if entityMap[tagProj] == 0 {
-				continue
-			}
-			tagProjIndex[tagProj] = partition.TagLocator{
-				FamilyOffset: i,
-				TagOffset:    j,
-			}
-		}
-	}
-	sidToIndex := make(map[common.SeriesID]int)
-	for idx, series := range seriesList {
-		sidToIndex[series.ID] = idx
-	}
-
-	return entityMap, tagSpecIndex, tagProjIndex, sidToIndex
-}
