+++ pkg/query/executor/interface.go
@@ -30,9 +30,9 @@ import (
 
 // StreamExecutionContext allows retrieving data through the stream module.
 type StreamExecutionContext interface {
-	Query(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamResultPuller, error)
-	Sort(ctx context.Context, opts pbv1.StreamSortOptions) (pbv1.StreamSortResult, error)
-	Filter(ctx context.Context, opts pbv1.StreamFilterOptions) (pbv1.StreamResultPuller, error)
+	Query(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error)
+	Sort(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamSortResult, error)
+	Filter(ctx context.Context, opts pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error)
 }
 
 // StreamExecutionContextKey is the key of stream execution context in context.Context.
